Rename the helper library to 'helper.js' and move it to src, so it's babelified.
