+++ src/Comments.jsx
@@ -7,7 +7,7 @@ import '../style/index.css';
 import { TextField, shapes } from '@cimpress/react-components';
 import CommentsClient from './CommentsClient';
 import { SERVICE_URL } from '../config';
-import { getSubFromJWT } from '../lib';
+import { getSubFromJWT } from './helper';
 
 let {Spinner} = shapes;
 
@@ -143,7 +143,7 @@ export default class _Comments extends React.Component {
     } else if (this.state.commentsIds.length > 0) {
       comments = this.state.commentsIds.map(
         (id, index) => <Comment className={'comment ' + ((index % 2 === 0) ? 'comment-even' : 'comment-odd')} key={id} accessToken={this.props.accessToken}
-                                commentUri={url + id} comment={this.state.commentObjects[id]} editComments={this.props.editComments}/>);
+          commentUri={url + id} comment={this.state.commentObjects[id]} editComments={this.props.editComments}/>);
     } else if (this.state.loading) {
       comments = <div>
         <div className="inline-spinner"><Spinner size={20}/></div>
